+++ internal/service/relation_service/follow_action.go
@@ -18,17 +18,17 @@ var (
 	ErrIvdFolUsr = errors.New("user not exist")
 )
 
-func FollowAction(hostID, guestId, actionType uint) error {
+func FollowAction(hostID, guestId, actionType int64) error {
 	return NewPostFollowActionFlow(hostID, guestId, actionType).Do()
 }
 
 type PostFollowActionFlow struct {
-	userId     uint
-	userToId   uint
-	actionType uint
+	userId     int64
+	userToId   int64
+	actionType int64
 }
 
-func NewPostFollowActionFlow(userId, userToId, actionType uint) *PostFollowActionFlow {
+func NewPostFollowActionFlow(userId, userToId, actionType int64) *PostFollowActionFlow {
 	return &PostFollowActionFlow{userId: userId, userToId: userToId, actionType: actionType}
 }
